+++ sphinx_astropy/conf/v1.py
@@ -126,6 +126,7 @@ suppress_warnings = ['app.add_directive', ]
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    'sphinx_astropy.ext.intersphinx_toggle',
     'sphinx.ext.autodoc',
     'sphinx.ext.intersphinx',
     'sphinx.ext.todo',
@@ -136,8 +137,8 @@ extensions = [
     'sphinx_automodapi.automodapi',
     'sphinx_automodapi.smart_resolver',
     'sphinx_astropy.ext.doctest',
-    'sphinx_astropy.ext.changelog_links']
-
+    'sphinx_astropy.ext.changelog_links',
+    'sphinx_astropy.ext.missing_static']
 
 if not on_rtd and LooseVersion(sphinx.__version__) < LooseVersion('1.4'):
     extensions.append('sphinx.ext.pngmath')